fixes bug in setparams + new tests
The bug is arguably in inspect.getmodule, and the changes here are really a (fairly weak) workaround.  The root of the problem is that inspect.getmodule does not take into account that the current directory may have change from the time the program was started.  In some cases, this makes it unable to locate the file for the __main__ module.  Nonsense ensues.