archival: src/onetime/munge.py
This is the script I used to generate the data for JM on 12-10-09
(actually 12-10-10, it turns out).  I'm committing this script only
for archival purposes; it is not expected to run as is, since it
requires some general utility modules that I have not yet committed
(these utility modules are fairly general, entirely independent of the
hmslincs data, so their absence should not be an obstacle to
understanding what onetime/munge.py is doing); also, the script
mentions all the data files it used.  (I will commit these files too,
as soon as I have a chance.)