finish implementation of withlimits...
...option for do_scatterplots.py.  Now, prefixing the execution of the
script with WITHLIMITS=1 will instruct the script to compute plot
limits common to all the plots; by default (or with WITHLIMITS=),
these limits are not computed (None is passed as the value of the lims
keyword in the call to write_scatterplot).

I updated README_121024W.txt to reflect this new usage.  (Actually,
the shell script src/do_scatterplots.sh committed earlier already
included this new usage.)

Also, this version of do_scatterplot now uses minmax.minmax (in
_range).