fixed picks_for_slider.tsv...
...(in sampledata/MH) so that the response time-average plot is
included in the same batch as the ones for the slider proper, and
therefore "contributes" to the computation of the common limits.

The invocation of do_scatterplots.py for this case should be changed
to

  WITHLIMITS=1 OUTPUTDIR=django/responses/static/responses/img/slider \
    python src/do_scatterplots.py sampledata/MH/picks_for_slider.tsv

Note the setting for the OUTPUTDIR global.  This setting is necessary
because otherwise the image called pakt,igf-1,100__perk,fgf1,100.png
would clobber an existing image with the same name but generated with
lims=None.