move javascript after css in head
trying to fix the issue where the javascript runs and positions the popups
before the css has loaded and positioned the imagemap properly. without
pathway.css the popups all render as visible blocks, pushing the imagemap down
beyond 4000px or so. the popups all get positioned off of that top coordinate,
leading to disaster.