Defend against broken libMagickWand on orchestra debian
wand.image won't even import on the debian systems as libMagickWand is one
version too old and is missing a critical symbol that the python module is
expecting. This commit moves the want import within the code that is only run
if the "no images" flag (-n) isn't passed, which means the -n flag will let
the affected scripts run their HTML generation parts on orchestra (which is
critical as those parts need to write to the DB).